Refactor imports to use absolute paths in code2seq module

+++ common.py
@@ -12,7 +12,7 @@ class ContextInfo:
         token1: str,
         path: str,
         token2: str,
-        attention_score: np.float,
+        attention_score: float,
         vector: np.ndarray,
     ) -> None:
         self._token1 = token1
@@ -22,7 +22,7 @@ class ContextInfo:
         self._vector = vector
 
     @property
-    def attention_score(self) -> np.float:
+    def attention_score(self) -> float:
         return self._attention_score
 
     @property
